+++ statuscheck/services/bitbucket.py
@@ -1,0 +1,6 @@
+from statuscheck.services._statuspage import BaseStatusPageAPI
+
+
+class ServiceAPI(BaseStatusPageAPI):
+    domain_key = 'bqlf8qjztdtr'
+    status_url = 'https://status.bitbucket.org/'