+++ statuscheck/services/cloudflare.py
@@ -1,0 +1,6 @@
+from statuscheck.services._statuspage import BaseStatusPageAPI
+
+
+class ServiceAPI(BaseStatusPageAPI):
+    domain_key = 'yh6f0r4529hb'
+    status_url = 'https://www.cloudflarestatus.com/'