+++ statuscheck/services/digitalocean.py
@@ -1,0 +1,6 @@
+from statuscheck.services._statuspage import BaseStatusPageAPI
+
+
+class ServiceAPI(BaseStatusPageAPI):
+    domain_key = 's2k7tnzlhrpw'
+    status_url = 'https://status.digitalocean.com/'