+++ statuscheck/services/mailgun.py
@@ -1,0 +1,6 @@
+from statuscheck.services._statuspage import BaseStatusPageAPI
+
+
+class ServiceAPI(BaseStatusPageAPI):
+    domain_key = '6jp439mdyy0k'
+    status_url = 'https://status.mailgun.com/'