+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.1.0] - 2019-01-20
+### Added
+- Bitbucket status checker
+- Cloudflare status checker
+- DigitalOcean status checker
+- Mailgun status checker
+
 ## [1.0.2] - 2019-01-15
 ### Changed
 - Fix Travis credentials mess caused by .org and .com domains